Wav audio

+++ AudioCompression/src/audioCompression/types/RawAudio.java
@@ -10,9 +10,9 @@ public interface RawAudio extends AudioCompressionType {
 	
 	public long getSamplesPerWindow();
 	
-	public long getNWindows();
+	public int getNWindows();
 	
-	public float getWindowOverlap();
+	public int getWindowOverlap();
 	
 	public int getNChannels();
 	
@@ -23,6 +23,11 @@ public interface RawAudio extends AudioCompressionType {
 	 */
 	public float[][] getAudioBuffer(int nsamples);
 	
+	
+	/**
+	 * window returned with dimensions [nChannels][nWindows][nSamples]
+	 * @return
+	 */
 	public float[][][] getAllWindows();
